requirements.txt
alembic==0.9.6
banal==0.3.3
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
dataset==1.0.5
decorator==4.2.1
feedparser==5.2.1
idna==2.6
jedi==0.11.1
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
normality==0.5.7
parso==0.1.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
Pygments==2.2.0
python-dateutil==2.6.1
python-editor==1.0.3
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
slackclient==1.1.0
SQLAlchemy==1.2.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.46.0
